# internal/api/api.swagger.yaml
openapi: 3.0.3
info:
  title: Online Courses Platform API
  description: API для платформы онлайн-курсов (аналог Docusaurus + LMS функционал)
  version: 0.1.0
  contact:
    name: Artem
    email: artem@example.com
servers:
  - url: https://api.courses-platform.com/v1
    description: Production server
  - url: http://localhost:3001
    description: Local development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Course:
      type: object
      properties:
        id:
          type: integer
          format: int64
        slug:
          type: string
        title:
          type: string
        subtitle:
          type: string
        description:
          type: string
        coverUrl:
          type: string
        status:
          type: string
          enum: [draft, published, archived]
        price:
          type: number
        currency:
          type: string
          example: EUR
        level:
          type: string
          enum: [beginner, intermediate, advanced]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CourseCreate:
      type: object
      required: [title, slug]
      properties:
        title:
          type: string
          minLength: 3
        slug:
          type: string
          pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
        subtitle:
          type: string
        description:
          type: string
        coverUrl:
          type: string
          format: uri
        price:
          type: number
          minimum: 0
        currency:
          type: string
          example: EUR
        level:
          type: string
          enum: [beginner, intermediate, advanced]
    CourseUpdate:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        subtitle:
          type: string
        description:
          type: string
        coverUrl:
          type: string
          format: uri
        status:
          type: string
          enum: [draft, published, archived]
        price:
          type: number
          minimum: 0
        currency:
          type: string
        level:
          type: string
          enum: [beginner, intermediate, advanced]
...
    Lesson:
      type: object
      properties:
        id:
          type: integer
          format: int64
        sectionId:
          type: integer
          format: int64
        title:
          type: string
        type:
          type: string
          enum: [video, text, quiz, assignment, pdf, coding, embed]
        content:
          type: string
          description: URL видео, текст, markdown, JSON для quiz и т.д.
        order:
          type: integer
          description: Порядковый номер урока внутри раздела
        durationSec:
          type: integer
          nullable: true
        isPublished:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    LessonCreate:
      type: object
      required: [title, type, order]
      properties:
        title:
          type: string
        type:
          type: string
          enum: [video, text, quiz, assignment, pdf, coding, embed]
        content:
          type: string
        order:
          type: integer
        durationSec:
          type: integer
        isPublished:
          type: boolean
          default: false
    LessonUpdate:
      type: object
      properties:
        title:
          type: string
        type:
          type: string
          enum: [video, text, quiz, assignment, pdf, coding, embed]
        content:
          type: string
        order:
          type: integer
        durationSec:
          type: integer
        isPublished:
          type: boolean
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        status:
          type: integer
paths:
...
  /courses:
    get:
      operationId: getCourses
      summary: Список всех опубликованных курсов (для всех пользователей)
      tags: [Courses]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 12
        - name: category
          in: query
          schema:
            type: string
        - name: level
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
          description: Поиск по названию или описанию
      responses:
        "200":
          description: Список курсов
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"

        "400":
          description: Некорректные параметры запроса

    post:
      operationId: createCourse
      summary: Создание нового курса (только для преподавателей и админов)
      tags: [Courses]

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseCreate"
      responses:
        "201":
          description: Курс успешно создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
        "400":
          description: Некорректные данные (например, slug уже занят)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Недостаточно прав
  /courses/{courseID}:
    get:
      operationId: getCourseByID
      summary: Получить подробную информацию о курсе
      tags: [Courses]
      parameters:
        - name: courseID
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Детали курса
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
        "404":
          description: Курс не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateCourse
      summary: Частично обновить курс (для преподавателей и админов)
      tags: [Courses]

      parameters:
        - name: courseID
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseUpdate"
      responses:
        "200":
          description: Курс обновлён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
        "400":
          description: Некорректные данные
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Нет прав на редактирование этого курса
        "404":
          description: Курс не найден

    delete:
      operationId: deleteCourse
      summary: Удалить курс (только для админов или владельца)
      tags: [Courses]

      parameters:
        - name: courseID
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Курс успешно удалён
        "403":
          description: Нет прав на удаление (например, курс имеет активных студентов)
        "404":
          description: Курс не найден
...
  /courses/{courseID}/sections/{sectionID}/lessons:
    get:
      operationId: getLessons
      summary: Получить список всех уроков в разделе курса
      tags: [Lessons]

      parameters:
        - name: courseID
          in: path
          required: true
          schema:
            type: string
        - name: sectionID
          in: path
          required: true
          schema:
            type: string
        - name: publishedOnly
          in: query
          schema:
            type: boolean
            default: true
          description: Показывать только опубликованные уроки (для студентов)
      responses:
        "200":
          description: Список уроков раздела
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lesson"
        "404":
          description: Курс или раздел не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      operationId: createLesson
      summary: Создать новый урок в разделе (для преподавателей/админов)
      tags: [Lessons]

      parameters:
        - name: courseID
          in: path
          required: true
          schema:
            type: string
        - name: sectionID
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LessonCreate"
      responses:
        "201":
          description: Урок успешно создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
        "400":
          description: Некорректные данные
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Нет прав на создание урока в этом курсе
        "404":
          description: Курс или раздел не найден
  /courses/{courseID}/sections/{sectionID}/lessons/{lessonID}:
    get:
      operationId: getLessonByID
      summary: Получить информацию об одном уроке
      tags: [Lessons]

      parameters:
        - name: courseID
          in: path
          required: true
          schema:
            type: string
        - name: sectionID
          in: path
          required: true
          schema:
            type: string
        - name: lessonID
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Детали урока
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
        "403":
          description: Нет доступа (например, урок не опубликован и пользователь не автор курса)
        "404":
          description: Урок не найден

    patch:
      operationId: updateLesson
      summary: Обновить урок (для преподавателей/админов)
      tags: [Lessons]

      parameters:
        - name: courseID
          in: path
          required: true
          schema:
            type: string
        - name: sectionID
          in: path
          required: true
          schema:
            type: string
        - name: lessonID
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LessonUpdate"
      responses:
        "200":
          description: Урок обновлён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
        "400":
          description: Некорректные данные
        "403":
          description: Нет прав на редактирование
        "404":
          description: Урок не найден

    delete:
      operationId: deleteLesson
      summary: Удалить урок (для преподавателей/админов)
      tags: [Lessons]

      parameters:
        - name: courseID
          in: path
          required: true
          schema:
            type: string
        - name: sectionID
          in: path
          required: true
          schema:
            type: string
        - name: lessonID
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Урок успешно удалён
        "403":
          description: Нет прав на удаление
        "404":
          description: Урок не найден
...